--- src/controller/loginController.js
@@ -1,157 +1,0 @@
-const bcrypt = require('bcryptjs');// Añade la importación de bcrypt
-const connection = require('../conexion');
-
-
-
-const getIndex = (req, res) => {
-    res.render('index');
-};
-
-const login = (req, res) => {
-    const username = req.body.username;
-    const password = req.body.contraseña;
-    const query = 'SELECT * FROM usuarios WHERE username = ?';
-   
-    connection.query(query, [username], (err, results) => {
-        if (err) throw err;
-        if (results.length > 0) {
-            // Verificar si la contraseña coincide utilizando bcrypt.compare()
-            const validPassword = bcrypt.compare(password, results[0].contraseña);
-                if (validPassword) {
-                    req.session.loggedIn = true; // Establece la sesión como iniciada
-                    req.session.username = username; // Guarda el nombre de usuario en la sesión
-                    res.redirect('/dashboard'); // Redirige al usuario al panel de control (dashboard)
-                } else {
-                    res.render('index', { mensaje: 'Contraseña inválida' });
-                }
-   
-        } else {
-            res.render('index', { mensaje: 'Usuario no valido' }); // Redirige al usuario de vuelta al formulario de inicio de sesión con un mensaje de error
-        }
-    
-    });
-};
-
-
-const logout = (req, res) => {
-    req.session.destroy((err) => {
-        if (err) {
-            console.error('Error destroying session:', err);
-        }
-        
-        res.redirect('/');
-    });
-};
-
-const getDashboard = (req, res) => {
-    if (req.session.loggedIn) {
-     res.render('dashboard', { username: req.session.username });
- } else {
-     res.render('error');
- }
- };
-
- const postRegister = async (req, res) => {
-    //const url = req.body;
-    const { name, apellido, username, password, fecha_nacimiento, phone } = req.body;
-    // Encriptar la contraseña
-    const hashedPassword = await bcrypt.hash(password, 10);
-    const insertQuery ='INSERT INTO usuarios (nombre, apellido, username, contraseña, fecha_nacimiento, telefono) VALUES (?, ?, ?, ?, ?, ?)';
-    const values = [name, apellido, username, hashedPassword, fecha_nacimiento, phone];
-    // Ejecutar la consulta INSERT
-    connection.query(insertQuery, values, function(error, results, fields) {
-    if (error) {
-        console.error('Error al insertar usuario:', error);
-        return;
-    }
-    //console.log('Usuario insertado correctamente');
-    res.render('index',{ mensaje: 'Usuario registrado' });
-    });
-};
-
- 
- const postLogin =(req, res) => {
-    const username = req.body.username;
-    const password = req.body.contraseña;
-    const query = 'SELECT * FROM usuarios WHERE username = ?';
-   
-    connection.query(query, [username], (err, results) => {
-        if (err) throw err;
-        if (results.length > 0) {
-            // Verificar si la contraseña coincide utilizando bcrypt.compare()
-            const validPassword = bcrypt.compare(password, results[0].contraseña);
-                if (validPassword) {
-                    req.session.loggedIn = true; // Establece la sesión como iniciada
-                    req.session.username = username; // Guarda el nombre de usuario en la sesión
-                    res.redirect('/dashboard'); // Redirige al usuario al panel de control (dashboard)
-                } else {
-                    res.render('index', { mensaje: 'Contraseña inválida' });
-                }
-   
-        } else {
-            res.render('index', { mensaje: 'Usuario no valido' }); // Redirige al usuario de vuelta al formulario de inicio de sesión con un mensaje de error
-        }
-    
-    });
-};
-
-const getRegistro = (req, res) => {
-    res.render('register');
-};
-
-const getContact = (req, res) => {
-    res.render('contact');
-};
-
-const getLogin = (req, res) => {
-    res.render('login');
-};
-
-/*
-
-const getComentarios = (req, res) => {
-    res.render('comentarios',{ username: req.session.username });
-};
-
-const getpublicaciones = (req, res) => {
-    res.render('publicaciones',{ username: req.session.username });
-};
-
-*/
-
-const getPerfil = (req, res) => {
-    const username = req.session.username;
-    const mensaje = req.query.mensaje;
-    // Consulta SQL para obtener los datos del perfil del usuario
-    const sql = 'SELECT * FROM usuarios WHERE username = ?';
-    connection.query(sql, [username], (err, results) => {
-        if (err) {
-            console.error('Error al obtener datos del perfil:', err);
-            res.status(500).send('Error al obtener datos del perfil');
-            return;
-        }
-
-        // Comprobar si se encontraron resultados
-        if (results.length > 0) {
-            // Renderizar la plantilla 'perfil' con los datos del usuario
-            //console.log(results[0].fecha_nacimiento); 
-            res.render('perfil', { username, usuario: results[0], mensaje});
-        } else {
-            res.status(404).send('Usuario no encontrado');
-        }
-    });
-};
-
-// Exporta las funciones y el enrutador
-module.exports = {
-    getIndex,
-    login,
-    logout,
-    getDashboard,
-    postLogin,
-    getRegistro,
-    postRegister,
-    getPerfil,
-    getLogin,
-    getContact
-};
