--- src/css/preorders.css
@@ -1,230 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body{
-    text-align: center;
-    font-family: sans-serif;
-    
-}
-
-.logo{
-    display: flex;
-    margin-left: 1.87em;
-}
-
-.logo img{
-    width:  3.1em;
-    height: 3.1em;
-}
-
-.logo2{
-    display: flex;
-    margin-right: 1.87em;
-}
-
-.logo2 img{
-    width: 2.5em;
-    margin-right: 0.31em;
-}
-
-
-.head{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    height: 4.37em;
-    background-color: #29c224;
-    position: fixed;
-    width: 100%;
-    z-index: 100%;
-    font-weight: bold; 
-    text-transform: uppercase;  /*Transforma en mayuscula*/
-      
-}
-
-.banner{
-    height: 60vh !important;
-    background: url(/img/banner.png) no-repeat center;  
-    background-attachment: fixed;   /*Para mover la foto scroll*/
-    background-size:cover;      /*cover: ocupa todo el ancho la imagen*/
-}  
-
-
-.menu2{
-    display: flex;
-    margin-right: 0.62em;
-}
-
-.menu2 a {
-    display: block;
-    padding: 1.25em 1.25em;
-    color: rgb(233, 231, 231);
-    text-decoration: none;
-    color: black;
-
-}
-
-.menu2 a:hover{
-    background-color: #b5ecad;
-}
-
-.logo2 img:hover{
-    background-color: #b5ecad;
-}
-
-.header {
-    display: flex;
-    align-items: center;
-}
-
-
-.header{
-    height: 30vh !important;
-    background: url(/img/banner.png);  
-    background-attachment: fixed;
-    background-size: contain;  /* para ajustar imagen en el contenedor */
-    background-repeat: no-repeat; /* Evita que la imagen se repita */
-}  
-
-.content{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.products-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 37em; 
-}
-.content-products{
-    text-align: center;
-    margin-top: auto;
-    
-}
-
-.page-footer{
-    border-bottom: 0.0625em solid #9c9a9a;
-    background-color: #eee7e7;
-    margin-top: 3.75em;
-
-}
-
-.tittle-footer{
-    margin-top: 0.3em;
-}
-
-
-.btn{
-    width: auto;
-    padding: 0.25em 0.625em;
-    font-size: 1rem;
-    line-height: 1.2rem;
-    color: aliceblue;
-    background-color: rgb(35, 126, 206);
-    margin-top: 1.25em;
-    cursor: pointer;
-    border-radius: 0.5em;
-    margin-left: 1.25em;
-}
-
-.subscribe{
-    display: flex;
-    min-height: 3.75em;
-    justify-content: center;
-    padding: 2.5em;
-    margin-bottom: 1.25em;
-    
-}
-
-.policy-text {
-    margin-top: 0.93em;
-    color: #000;
-    display: block;
-}
-
-.input-text{
-    width: 18.75em; /* Ancho del cuadro de texto */
-    height: 1.87em; /* Altura del cuadro de texto */
-    font-size: 1rem; 
-}
-
-
-.policy-link {
-    color: rgb(35, 126, 206); /* Color azul para los enlaces */
-    text-decoration: none; /* Remueve el subrayado de los enlaces */
-}
-
-.policy-link:hover {
-    text-decoration: underline; /* Añade subrayado al pasar el ratón sobre el enlace */
-}
-
-.footer-content{
-    display: flex;
-    list-style: none none;
-    justify-content: center;
-    text-decoration: none;
-    width: 100%;
-    height: 6.25em;  
-}
-
-.footer-content h4{
-    text-transform: uppercase;
-}
-
-.footer-content ul li{
-    list-style: none;
-}
-
-.about-footer{
-    padding-right: 3.125em;
-    
-}
-
-.listas-about li a{
-    text-decoration: none;
-    color: #9c9a9a;
-    text-align: left;
-}
-
-.listas-about{
-    margin-top: 0.1875em ;
-}
-
-.contacto-footer{
-    padding-left: 10em;
-   
-    
-}
-.social-icons{
-    display: flex;
-    width: 1.15em;
-    height: 1.1em;
-    margin: 0.1em;
-    flex-direction: row; /* Cambiar la dirección a vertical */
-    margin-top: 0.5em;
-    
-    
-}
-
-
-.vertical-border {
-    display: flex;
-    border-right: 0.09em solid #9c9a9a;
-    padding-left: 6.2em;
-    margin-top: -2.1em;
-    
-}
-
-.dropdown-icon{
-    display: none;
-}
-
-.points-icon{
-    display: none;
-}
