--- src/routes/image.js
@@ -1,14 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const multer = require('multer');
-const upload = multer();
-//
-const imageController = require('../controller/imageController');
-
-router
-    //login
-    .get('/imagen/:id',imageController.getImage)
-    .post('/modifica_foto',upload.single('image'), imageController.modificaFoto);
-    
-
-module.exports = router;
