+++ src/server.js
@@ -2,8 +2,7 @@ const express = require('express');
 const session = require('express-session');
 const bodyParser = require('body-parser');
 //importamos las rutas
-const loginRoutes = require('./routes/login');
-const imageRoutes = require('./routes/image');
+const loginRoutes = require('./routes/rutas');
 
 const app = express();
 const port = 3000;
@@ -25,7 +24,6 @@ app.set('views', './src/views');
 
 // Routes
 app.use('/', loginRoutes);
-app.use('/', imageRoutes);
 
 
 app.listen(port, () => {
