+++ src/views/index.ejs
@@ -21,7 +21,6 @@
                 <a href="/news">Novedades</a>
                 <a href="/offers">Ofertas</a>
                 <a href="/catalogo">Catalogo</a>
-                <a href="/preorders">Prepedidos</a> 
                 <a href="/contact">Contacto</a>
                 
             </nav>
@@ -30,7 +29,7 @@
                 <div class="dropdown-icon">
                     <img src="./images/desplegable.png" alt="Desplegable">
                 </div>
-                <a href="/register"><img src="./images/user.png" alt="icono para iniciar/registrarse"></a>
+                <a href="/login"><img src="./images/user.png" alt="icono para iniciar/registrarse"></a>
                 <a href="#"><img class="carro" src="./images/carro.png" alt="carro de compra"></a>
             </div>
